misc updates in runner package
