+++ internal/runner/banner.go
@@ -1,3 +1,4 @@
+// Package runner executes the enumeration process.
 package runner
 
 import (
