--- internal/runner/defaults.go
@@ -1,6 +1,0 @@
-package runner
-
-const (
-	// Default directory used to save protocols traffic
-	DefaultDumpTrafficOutputFolder = "output"
-)
