--- internal/runner/doc.go
@@ -1,2 +1,0 @@
-// Package runner executes the enumeration process.
-package runner
